Rename data loss detection namespace
